agents/report_agent/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
PyJWT>=2.8.0
httpx>=0.27.0
